masconLevelに基づいて燃料を消費するシステムを作成

+++ moorestech_server/Assets/Scripts/Game.Train/Unit/IFuelProviderTrainCarContainer.cs
@@ -10,6 +10,6 @@ namespace Game.Train.Unit
         ///     消費にかかる時間
         ///     Time required for consumption
         /// </returns>
-        float ConsumeFuel(TrainCar trainCar);
+        double ConsumeFuel(TrainCar trainCar);
     }
 }
